+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
+- Allow empty user data for registrations (#505)
 - Improve the code autoformatting (#502)
 
 ### Deprecated